re-designed the whole load image/action cancelling

+++ Hitomi-Scroll-Viewer/MainWindow.xaml.cs
@@ -15,12 +15,19 @@ namespace Hitomi_Scroll_Viewer {
         public static readonly string IMAGE_DIR = ROOT_DIR + @"\images";
         public static readonly string IMAGE_EXT = ".webp";
 
+        private readonly SearchPage _sp;
+        private readonly ImageWatchingPage _iwp;
         private readonly Page[] _appPages;
         private static int _currPageNum = 0;
+
         public AppWindow appWindow;
 
-        public static Gallery gallery;
-        public static List<Gallery> bmGalleries;
+        public Gallery gallery;
+        public List<Gallery> bmGalleries;
+
+        public CancellationTokenSource cts;
+        public readonly object actionLock = new();
+        public bool isInAction = false;
 
         public MainWindow() {
             InitializeComponent();
@@ -29,15 +36,15 @@ namespace Hitomi_Scroll_Viewer {
             Directory.CreateDirectory(ROOT_DIR);
             Directory.CreateDirectory(IMAGE_DIR);
 
-            SearchPage sp = new(this);
-            ImageWatchingPage iwp = new(this);
-            _appPages = new Page[] { sp, iwp };
+            _sp = new(this);
+            _iwp = new(this);
+            _appPages = new Page[] { _sp, _iwp };
 
-            sp.Init(iwp);
+            _sp.Init(_iwp);
 
             // Switch page on double click
             RootFrame.DoubleTapped += (object _, DoubleTappedRoutedEventArgs _) => {
-                iwp.SetAutoScroll(false);
+                _iwp.StartStopAutoScroll(false);
                 SwitchPage();
             };
 
@@ -52,8 +59,7 @@ namespace Hitomi_Scroll_Viewer {
 
             // Handle window close
             Closed += (object _, WindowEventArgs _) => {
-                iwp.DisableControls();
-                ImageWatchingPage.WaitActionFinish();
+                HandleClose();
                 if (gallery != null) {
                     if (!IsBookmarked()) {
                         DeleteGallery(gallery);
@@ -67,10 +73,6 @@ namespace Hitomi_Scroll_Viewer {
         public void SwitchPage() {
             _currPageNum = (_currPageNum + 1) % _appPages.Length;
             RootFrame.Content = _appPages[_currPageNum];
-        }
-
-        public Page GetPage() {
-            return _appPages[_currPageNum];
         }
 
         private readonly ContentDialog _dialog = new() {
@@ -90,7 +92,7 @@ namespace Hitomi_Scroll_Viewer {
             await _dialog.ShowAsync();
         }
 
-        public static bool IsBookmarked() {
+        public bool IsBookmarked() {
             for (int i = 0; i < bmGalleries.Count; i++) {
                 if (bmGalleries[i].id == gallery.id) {
                     return true;
@@ -99,15 +101,39 @@ namespace Hitomi_Scroll_Viewer {
             return false;
         }
 
-        public static bool IsBookmarkFull() {
+        public bool IsBookmarkFull() {
             return bmGalleries.Count == SearchPage.MAX_BOOKMARK_PAGE * SearchPage.MAX_BOOKMARK_PER_PAGE;
         }
 
         public static void DeleteGallery(Gallery removingGallery) {
             string path = IMAGE_DIR + @"\" + removingGallery.id;
-            if (Directory.Exists(path)) {
-                Directory.Delete(path, true);
+            if (Directory.Exists(path)) Directory.Delete(path, true);
+        }
+
+        public void StartAction(bool start) {
+            lock (actionLock) {
+                if (start) {
+                    isInAction = true;
+                    _iwp.LoadingControlBtn.Label = "Cancel Loading";
+                    _iwp.LoadingControlBtn.Icon = new SymbolIcon(Symbol.Cancel);
+                }
+                _sp.EnableControls(!start);
+                _iwp.EnableControls(!start);
+                if (!start) {
+                    _iwp.LoadingControlBtn.Label = "Reload Gallery";
+                    _iwp.LoadingControlBtn.Icon = new SymbolIcon(Symbol.Sync);
+                    isInAction = false;
+                }
             }
+        }
+
+        private async void HandleClose() {
+            _iwp.WaitBookmarking();
+            lock (actionLock) if (isInAction) cts.Cancel();
+            while (isInAction) {
+                await Task.Delay(10);
+            }
+            cts.Dispose();
         }
     }
 }
